:art: refactor: 길찾기 경로 계산 함수 분리
- calculatePath 함수를 src/app/utils/pathCalculator.ts 파일로 이동합니다.
