+++ scope-app/src/app/page.tsx
@@ -26,20 +26,7 @@ import { Vehicle, ParkingSpace, BuildingEntrance, ParkingFloor, ParkingLot, Stat
 
 import { mockVehicles, mockParkingLots } from "./data/mockData";
 
-// --- 길찾기 경로 계산 함수 ---
-const calculatePath = (start: { x: number; y: number }, end: { x: number; y: number }): { x: number; y: number }[] => {
-  const path = [];
-  const steps = 20;
-
-  for (let i = 0; i <= steps; i++) {
-    const t = i / steps;
-    const x = start.x + (end.x - start.x) * t;
-    const y = start.y + (end.y - start.y) * t + Math.sin(t * Math.PI) * 20;
-    path.push({ x, y });
-  }
-
-  return path;
-};
+import { calculatePath } from "./utils/pathCalculator";
 
 // --- 컴포넌트 Props 타입 정의 ---
 interface StatsCardProps {
